// src/types.ts
export type JsonReadOptions =
  | {
      encoding?: BufferEncoding | null;
      flag?: string;
      throws?: boolean;
      // eslint-disable-next-line @typescript-eslint/consistent-type-imports
      fs?: typeof import("fs");
      reviver?: (key: any, value: any) => any;
    }
  | BufferEncoding
  | null
  | undefined;

export type JsonWriteOptions =
  | {
      encoding?: BufferEncoding | null;
      mode?: string | number;
      flag?: string;
      // eslint-disable-next-line @typescript-eslint/consistent-type-imports
      fs?: typeof import("fs");
      EOL?: string;
      finalEOL?: boolean;
      spaces?: string | number;
      replacer?: (key: string, value: any) => any;
    }
  | BufferEncoding
  | null
  | undefined;